+++ example/android/app/src/main/kotlin/dev/dartoos/android_test/MainActivity.kt
@@ -1,0 +1,6 @@
+package dev.dartoos.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}